Move array to own file
